+++ main.h
@@ -14,7 +14,7 @@ int _putchar(char str);
 int printString(char *str);
 int printInt(int n);
 int print_binary(int n);
-void print_number(int n);
+void print_number(long int n);
 int number_nums(int n);
 int print_reverse(char *str);
 char rot13(char c);
